app/requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
Flask-RESTful==0.3.10
apispec==6.6.1
marshmallow==3.20.1
flask-apispec==0.11.4
Werkzeug==2.3.8
mindee ==4.12.0
gunicorn==20.1.0
pdfplumber
flask-cors